Completed second test
I spend a significant of time reading Array methods and trying them. But Googling helped in the end. I found a code in StackOverflow that worked in this scenario. I learned three Array methods well: map, filter and includes.

The Array.map() method takes a callbackFn as an argument and iterates over all elements, processing them to generate a new array.

The Array.filter() method is a bit different in that, it gives back a portion of elements that meet the provided criteria run in the callbackFn.

The Array.includes() method takes in an argument and checks whether that argument matches any of the elements in the array. If yes, it returns TRUE. So, in our case, we joined filter and includes together.

This is how filter and includes work together in this code:
The includes receives the x value from the inputArray one at a time, and for each x, it quickly checks whether its removalValue array contains the x. If it does, it returns TRUE. Ideally, the filter function would identify the element that got evaluated as TRUE and begin putting it in the intersection array. BUT, as per our requirement, we want the elements that don't match, i.e. the ones that give FALSE. Hence, we have added the exclamantion notation to evaluate false.