Complete all tests
The leap year calculation instruction was unnecessarily complicated. The struggle to understand the parameters of calculating leap years made me write wrong code after the first test. It wasn't until I googled, I learned about the nested rules of the leap year calculation.

The solution code is quite elegant. My code is not that elegant but the learning was good.