Complete all tests.
The solution code is unnecessarily difficult. I googled the removal of punctuation marks and spaces and received the necessary regex code for replacement. The method I've undertaken, uses minimum amount of chained functions. Also, the solution code seems smart, but I don't know how the code would work if the input characters were non-english alphabets. This is not a commentary on comparision with the solution code, but since I just had a look at it, I'm thinking out loud.

This exercise was fun because it allowed to think in layers one after another.