Completed all tests
Setting up the code for the initial test was a bit challenging and the wikipedia formula initially threw me in unnecessary mathematician role. But shortly, I wrote pseudocode and things became clear.

The nature of requirement with a one place displaced index threw an error but solving that taught about paying attention and being present about the workings of the code.