Completed the exercise
Dealing with objects in the array was new for me. There seemed many ways to accomplish this but I wasn't sure.

Finally, googling helped. The solution of the exercise has only array.map() method. But as I learned from the online reference, adding filter() method is more useful in filtering out objects that do not have title in them.