Complete first test
This was the longest it took me to come to the right code to get the result.

In a debugger, I had the code return Ray as expected, but while jest testing, the test repeatedly returned undfined. It was frustrating. Finally, reading the test equality check, I realized, the test was comparing the return array's name key's value. So, instead of returning person.name, I passed person to result. This worked.