Completed all tests
While working on the third test, I realized that the second test was technically successful, but it had a major flaw in testing. Because the missing deathYear wasn't associated with the oldest person, the answer of the test matched with the expected test result. And that's why, the test designers made their point by providing a test that checks whether the code works if the oldest person, in this case Carly, gets returned.

Interestingly, my code didn't return the right answer. I realized that I was using the new Date() as it is. While the deathYear required only Year. So, I converted it to year and it worked.